Folder move, css adjustment for mobile, remove contact me whil ein dev (#77)
