+++ src/components/homecards/Cards.css
@@ -47,7 +47,7 @@ h1 {
 .cards__item__pic-wrap {
     position: relative;
     width: 100%;
-    padding-top: 67%;
+    padding-top: 90%;
     overflow: hidden;
 }
