+++ src/components/pages/Home.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import '../../App.css';
-import Cards from '../projects/Cards';
+import Cards from '../homecards/Cards';
 import HeroSection from '../home/HeroSection';
 import Footer from '../common/Footer';
 
@@ -10,7 +10,6 @@ function Home() {
     <>
       <HeroSection />
       <Cards />
-      <Footer />
     </>
   );
 }
